Completed How Many Points
Set up the templates in the way that I wanted the information I wanted it to be displayed. I then built the form up to contain what info I wanted it to, and how it should be properly handled. Finally I set up the controller which handled the connections, and making sure that the template could be reached.